+++ src/lib.rs
@@ -1,7 +1,11 @@
 //! SUPER Android Analyzer
 
+// Allowing these at least for now.
+#![allow(unknown_lints, missing_docs_in_private_items, print_stdout, stutter, option_unwrap_used,
+    result_unwrap_used, integer_arithmetic, cast_possible_truncation, cast_possible_wrap,
+    indexing_slicing, cast_precision_loss, cast_sign_loss)]
 #![forbid(deprecated, overflowing_literals, stable_features, trivial_casts, unconditional_recursion,
-    plugin_as_library, unused_allocation, trivial_numeric_casts, unused_features, while_truem,
+    plugin_as_library, unused_allocation, trivial_numeric_casts, unused_features, while_true,
     unused_parens, unused_comparisons, unused_extern_crates, unused_import_braces, unused_results,
     improper_ctypes, non_shorthand_field_patterns, private_no_mangle_fns, private_no_mangle_statics,
     filter_map, used_underscore_binding, option_map_unwrap_or, option_map_unwrap_or_else,
@@ -12,149 +16,98 @@
     pub_enum_variant_names, shadow_reuse, shadow_same, shadow_unrelated, similar_names,
     single_match_else, string_add, string_add_assign, unicode_not_nfc, unseparated_literal_suffix,
     use_debug, wrong_pub_self_convention, doc_markdown)]
-// Allowing these at least for now.
-#![allow(missing_docs_in_private_items, unknown_lints, print_stdout, stutter, option_unwrap_used,
-    result_unwrap_used, integer_arithmetic, cast_possible_truncation, cast_possible_wrap,
-    indexing_slicing, cast_precision_loss, cast_sign_loss)]
 
 
 #[macro_use]
-pub extern crate clap;
-pub extern crate colored;
-pub extern crate xml;
-pub extern crate serde;
-pub extern crate serde_json;
-pub extern crate chrono;
-pub extern crate toml;
-pub extern crate regex;
+extern crate clap;
+extern crate colored;
+extern crate xml;
+extern crate serde;
+extern crate serde_json;
 #[macro_use]
-pub extern crate lazy_static;
-pub extern crate rustc_serialize;
-pub extern crate open;
-pub extern crate bytecount;
-pub extern crate handlebars;
+extern crate serde_derive;
+extern crate chrono;
+extern crate toml;
+extern crate regex;
 #[macro_use]
-pub extern crate log;
-pub extern crate env_logger;
+extern crate lazy_static;
+extern crate rustc_serialize;
+extern crate open;
+extern crate bytecount;
+extern crate handlebars;
 #[macro_use]
-pub extern crate error_chain;
-pub extern crate abxml;
-pub extern crate md5;
-pub extern crate sha1;
-pub extern crate sha2;
+extern crate log;
+extern crate env_logger;
+#[macro_use]
+extern crate error_chain;
+extern crate abxml;
+extern crate md5;
+extern crate sha1;
+extern crate sha2;
 
-pub mod criticality;
-pub mod error;
+mod error;
+/// Command Line Interface
 pub mod cli;
-pub mod decompilation;
-pub mod static_analysis;
-pub mod results;
-pub mod config;
-pub mod utils;
+mod decompilation;
+mod static_analysis;
+mod results;
+mod config;
+mod utils;
+mod criticality;
 
-pub use std::{fs, io, fmt, result};
-pub use std::path::Path;
-pub use std::fmt::Display;
-pub use std::str::FromStr;
-pub use std::error::Error as StdError;
-pub use std::io::Write;
-pub use std::time::{Instant, Duration};
-pub use std::thread::sleep;
-pub use std::collections::BTreeMap;
+use std::fs;
+use std::time::{Instant, Duration};
+use std::thread::sleep;
+use std::collections::BTreeMap;
+use std::path::{Path, PathBuf};
+use clap::ArgMatches;
 
-pub use serde::ser::{Serialize, Serializer};
-pub use colored::Colorize;
+use colored::Colorize;
 
-pub use log::{LogRecord, LogLevelFilter, LogLevel};
-pub use env_logger::LogBuilder;
-pub use std::env;
-pub use criticality::*;
-pub use decompilation::*;
-pub use static_analysis::*;
+use log::{LogRecord, LogLevelFilter, LogLevel};
+use env_logger::LogBuilder;
+use std::env;
+use decompilation::*;
+use static_analysis::*;
 pub use results::*;
 pub use error::*;
+// use criticality::*;
 pub use config::Config;
 pub use utils::*;
 
-pub static BANNER: &'static str = include_str!("banner.txt");
+pub static BANNER: &str = include_str!("banner.txt");
 
 
-pub fn run() -> Result<()> {
-    let cli = cli::generate().get_matches();
-    let verbose = cli.is_present("verbose");
-    initialize_logger(verbose);
+/// Initialize the config with the config files and command line options
+/// On UNIX, if local file ('config.toml') does not exists, but the global one does
+/// ('/etc/super-analyzer/config.toml'), the latter is used.
+/// Otherwise, the local file is used.
+/// Finally, if non of the files could be loaded, the default config is used
+pub fn initialize_config(cli: ArgMatches<'static>) -> Result<Config> {
+    let config_path = PathBuf::from("config.toml");
+    let global_config_path = PathBuf::from("/etc/super-analyzer/config.toml");
 
-    let mut config = match Config::from_cli(cli) {
-        Ok(c) => c,
-        Err(e) => {
-            print_warning(format!(
-                "There was an error when reading the config.toml file: {}",
-                e.description()
-            ));
-
+    let mut config =
+        if cfg!(target_family = "unix") && !config_path.exists() && global_config_path.exists() {
+            Config::from_file(&global_config_path).chain_err(|| {
+                format!("There was an error when reading the /etc/super-analyzer/config.toml file")
+            })?
+        } else if config_path.exists() {
+            Config::from_file(&PathBuf::from("config.toml")).chain_err(
+                || {
+                    format!("There was an error when reading the config.toml file")
+                },
+            )?
+        } else {
+            print_warning("Config file not found. Using default configuration");
             Config::default()
-        }
-    };
+        };
 
-    if !config.check() {
-        let mut error_string = String::from("Configuration errors were found:\n");
-        for error in config.get_errors() {
-            error_string.push_str(&error);
-            error_string.push('\n');
-        }
-        error_string.push_str(
-            "The configuration was loaded, in order, from the following files: \
-                               \n\t- Default built-in configuration\n",
-        );
-        for file in config.get_loaded_config_files() {
-            error_string.push_str(&format!("\t- {}\n", file.display()));
-        }
+    config.decorate_with_cli(cli).chain_err(
+        || "There was an error reading config from CLI",
+    )?;
 
-        return Err(ErrorKind::Config(error_string).into());
-    }
-
-    if config.is_verbose() {
-        for c in BANNER.chars() {
-            print!("{}", c);
-            io::stdout().flush().unwrap();
-            sleep(Duration::from_millis(3));
-        }
-        println!(
-            "Welcome to the SUPER Android Analyzer. We will now try to audit the given \
-                  application."
-        );
-        println!(
-            "You activated the verbose mode. {}",
-            "May Tux be with you!".bold()
-        );
-        println!();
-        sleep(Duration::from_millis(1250));
-    }
-
-    let mut benchmarks = BTreeMap::new();
-
-    let total_start = Instant::now();
-    for package in config.get_app_packages() {
-        config.reset_force();
-        analyze_package(package, &mut config, &mut benchmarks)
-            .chain_err(|| "Application analysis failed")?;
-    }
-
-    if config.is_bench() {
-        let total_time = Benchmark::new("Total time", total_start.elapsed());
-        println!();
-        println!("{}", "Benchmarks:".bold());
-        for (package_name, benchmarks) in benchmarks {
-            println!("{}:", package_name.italic());
-            for bench in benchmarks {
-                println!("{}", bench);
-            }
-            println!();
-        }
-        println!("{}", total_time);
-    }
-
-    Ok(())
+    Ok(config)
 }
 
 /// Analyzes the given package with the given config.
@@ -252,7 +205,12 @@ pub fn analyze_package<P: AsRef<Path>>(
 
     let report_start = Instant::now();
     results.generate_report(config, &package_name).chain_err(
-        || "There was an error generating the results report",
+        || {
+            format!(
+                "There was an error generating the results report. Tried to generate at: {}",
+                config.get_results_folder().join(&package_name).display()
+            )
+        },
     )?;
 
     if config.is_verbose() {
@@ -378,7 +336,7 @@ pub fn initialize_logger(is_verbose: bool) {
 }
 
 #[cfg(test)]
-pub mod tests {
+mod tests {
     use Criticality;
     use std::str::FromStr;
